+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Version latest (0.1.7?, 2019-09-??)
 
 - ...
+
+
+Version 0.3.1 (2019-09-16)
+
 - enh: add documentation in readme about recommendations
 - enh: stricter "overused" classifier, lambda classifier to output a 2nd classification of hourly vs daily resolution